test: complete trustees processes tests
