+++ decidim-elections/spec/spec_helper.rb
@@ -10,12 +10,3 @@ require "decidim/dev/test/base_spec_helper"
 require "decidim/forms/test"
 
 WebMock.disable_net_connect!(allow_localhost: true, allow: Decidim::BulletinBoard.bulletin_board_server)
-
-def in_browser(name)
-  old_session = Capybara.session_name
-
-  Capybara.session_name = name
-  yield
-
-  Capybara.session_name = old_session
-end
