+++ decidim-elections/spec/system/admin/admin_manages_election_steps_spec.rb
@@ -3,25 +3,15 @@
 require "spec_helper"
 
 describe "Admin manages election steps", :slow, type: :system do
-  let(:manifest_name) { "elections" }
-
-  include_context "when managing a component as an admin" do
-    let(:admin_component_organization_traits) { [:secure_context] }
-  end
-
-  before do
-    election
-    login_as user, scope: :user
-    visit_component_admin
-  end
+  include Decidim::Elections::FullElectionHelpers
+  include_context "with test bulletin board"
+  include_context "when admin manages elections"
 
   describe "setup an election" do
-    let!(:election) { create :election, :ready_for_setup, component: current_component, title: { en: "English title", es: "" } }
+    let(:election) { create :election, :ready_for_setup, component: current_component, title: { en: "English title", es: "" } }
 
     it "performs the action successfully" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within "form.create_election" do
         expect(page).to have_content("The election has at least 1 question.")
@@ -45,14 +35,10 @@ describe "Admin manages election steps", :slow, type: :system do
   end
 
   describe "start the key ceremony" do
-    include_context "with test bulletin board"
-
-    let!(:election) { create :election, :bb_test, :created, component: current_component }
+    let(:election) { create :election, :bb_test, :created, component: current_component }
 
     it "performs the action successfully" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within ".form.created" do
         expect(page).to have_content("Trustees")
@@ -72,29 +58,36 @@ describe "Admin manages election steps", :slow, type: :system do
     end
   end
 
-  describe "view key ceremony step" do
-    let!(:election) { create :election, :key_ceremony, component: current_component }
+  describe "view key ceremony step", :slow, download: true do
+    include_context "when performing the whole process"
 
     it "shows the step information" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      setup_election
 
-      within ".content.key_ceremony" do
-        expect(page).to have_content("Key ceremony")
-      end
+      visit_steps_page
+      expect(page).to have_content("Key ceremony")
+      expect(page).to have_css(".loading-spinner") # It shows the loading icon
+      expect(page).not_to have_css("svg.icon--task") # The trustees didn't participate yet
+      expect(page).to have_link("Continue", class: "disabled")
+
+      download_election_keys(0)
+      download_election_keys(1)
+      download_election_keys(2)
+
+      visit_steps_page
+      expect(page).to have_content("Key ceremony")
+      expect(page).not_to have_css(".loading-spinner") # It's not waiting for any trustee
+      expect(page).to have_css("svg.icon--task") # All the trustees are active
+      expect(page).not_to have_link("Continue", class: "disabled")
+      expect(page).to have_link("Continue")
     end
   end
 
   describe "start the voting period" do
-    include_context "with test bulletin board"
-
-    let!(:election) { create :election, :bb_test, :key_ceremony_ended, component: current_component }
+    let(:election) { create :election, :bb_test, :key_ceremony_ended, component: current_component }
 
     it "performs the action successfully" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within ".form.key_ceremony_ended" do
         expect(page).to have_content("The election will start soon.")
@@ -115,15 +108,11 @@ describe "Admin manages election steps", :slow, type: :system do
   end
 
   describe "voting period" do
-    include_context "with test bulletin board"
-
-    let!(:election) { create :election, :bb_test, :vote, component: current_component }
+    let(:election) { create :election, :bb_test, :vote, component: current_component }
 
     context "with no vote statistics" do
       it "shows text about vote statistics" do
-        within find("tr", text: translated(election.title)) do
-          page.find(".action-icon--manage-steps").click
-        end
+        visit_steps_page
 
         within "#vote-stats" do
           expect(page).to have_content("Vote Statistics")
@@ -139,9 +128,7 @@ describe "Admin manages election steps", :slow, type: :system do
       let!(:user_2_votes) { create :vote, election: election, status: "accepted", voter_id: "voter_#{user_2.id}" }
 
       it "shows votes and unique voters" do
-        within find("tr", text: translated(election.title)) do
-          page.find(".action-icon--manage-steps").click
-        end
+        visit_steps_page
 
         within "#vote-stats" do
           expect(page).to have_content("Votes")
@@ -158,14 +145,10 @@ describe "Admin manages election steps", :slow, type: :system do
   end
 
   describe "end the voting period" do
-    include_context "with test bulletin board"
-
-    let!(:election) { create :election, :bb_test, :vote, :finished, component: current_component }
+    let(:election) { create :election, :bb_test, :vote, :finished, component: current_component }
 
     it "performs the action successfully" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within ".form.vote" do
         expect(page).to have_content("The election has ended.")
@@ -186,14 +169,10 @@ describe "Admin manages election steps", :slow, type: :system do
   end
 
   describe "start the tally" do
-    include_context "with test bulletin board"
-
-    let!(:election) { create :election, :bb_test, :vote_ended, component: current_component }
+    let(:election) { create :election, :bb_test, :vote_ended, component: current_component }
 
     it "performs the action successfully" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within ".form.vote_ended" do
         expect(page).to have_content("Vote period ended")
@@ -214,34 +193,38 @@ describe "Admin manages election steps", :slow, type: :system do
   end
 
   describe "report missing trustee" do
-    let!(:election) { create :election, :tally, component: current_component }
+    let(:election) { create :election, :bb_test, :tally, component: current_component }
     let(:trustee) { election.trustees.first }
 
     it "marks the trustee as missing" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
+
+      # allows admin to mark trustees as missing
+      expect(page).to have_selector("button", text: "Mark as missing")
 
       within find("tr", text: trustee.name) do
         click_button "Mark as missing"
       end
 
-      within "#trustees_process" do
-        expect(page).to have_css("svg.icon--task")
-        expect(page).not_to have_content("Mark as missing")
+      expect(page).to have_admin_callout("successfully")
+
+      # shows the trustee as missing
+      within find("tr", text: trustee.name) do
+        expect(page).to have_css("svg.icon--x")
       end
+
+      # don't allow to mark more trustees as missing
+      expect(page).not_to have_selector("button", text: "Mark as missing")
     end
   end
 
   describe "tally ended" do
-    let!(:election) { create :election, :tally_ended, component: current_component }
-    let!(:question) { election.questions.first }
-    let!(:answer) { question.answers.first }
+    let(:election) { create :election, :tally_ended, component: current_component }
+    let(:question) { election.questions.first }
+    let(:answer) { question.answers.first }
 
     it "shows the calculated results" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within ".form.tally_ended" do
         expect(page).to have_content("Calculated results")
@@ -253,16 +236,12 @@ describe "Admin manages election steps", :slow, type: :system do
   end
 
   describe "publishing results" do
-    include_context "with test bulletin board"
-
-    let!(:election) { create :election, :bb_test, :tally_ended, component: current_component }
-    let!(:question) { election.questions.first }
-    let!(:answer) { question.answers.first }
+    let(:election) { create :election, :bb_test, :tally_ended, component: current_component }
+    let(:question) { election.questions.first }
+    let(:answer) { question.answers.first }
 
     it "performs the action successfully" do
-      within find("tr", text: translated(election.title)) do
-        page.find(".action-icon--manage-steps").click
-      end
+      visit_steps_page
 
       within ".form.tally_ended" do
         expect(page).to have_content("Calculated results")
@@ -279,6 +258,17 @@ describe "Admin manages election steps", :slow, type: :system do
       within ".content.results_published" do
         expect(page).to have_content("Results published")
       end
+    end
+  end
+
+  def visit_steps_page
+    election
+
+    relogin_as user, scope: :user
+    visit_component_admin
+
+    within find("tr", text: translated(election.title)) do
+      page.find(".action-icon--manage-steps").click
     end
   end
 end
